make checkmasses a rp command
